+++ template_variables/RELEASES.md
@@ -235,3 +235,11 @@ Based on:
 - Speakeasy CLI 1.453.1 (2.474.4) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.12.3] template_variables
+
+## 2024-12-10 00:31:45
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/template-variables.yaml
+- Speakeasy CLI 1.453.8 (2.474.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.12.4] template_variables